Import Landlock Config
The goal of Landlock Config is to provide a simple and flexible
configuration format for defining security policies using Landlock on
any kernel versions.  This format aims to be user-friendly, allowing for
manual creation or generation through JSON libraries.  It can be used as
an intermediate representation leveraged by higher-level configuration
formats or languages.  The associated library is designed to be
embeddable in various programs, facilitating sandboxing and providing
well-defined security and compatibility guarantees.

Signed-off-by: Mickaël Salaün <mic@digikod.net>