Add access right groups
Add groups for the filesystem and the network access rights.

The filesystem gets (from v1 to v5):
- vX.all
- vX.read_execute
- vX.read_write

The network gets (from v4 to v5):
- vX.all

Groups are only defined starting with an ABI version that supports a
class of access rigths.  For instance, network groups only start with
v4.

Update the example configurations with a minimal one and the same but
with explicit access rights.

Signed-off-by: Mickaël Salaün <mic@digikod.net>