Remove support for file descriptor in the configuration
File descriptor are always relative to the caller, so it does not make a
lot of sense to specify them in a configuration file.  Additionally,
there is a good chance for these file descriptors to be misused or just
not used at all.  If we need to rely on file descriptors, we should use
a Landlock library.

Rename the "parentFd" field to "parent" because it would not make sense
to keep the "Fd" part.  Only accept file paths in this "parent" field.

Simplify the build_ruleset() signature.

Signed-off-by: Mickaël Salaün <mic@digikod.net>