Add C/C++ FFI bindings
TODO: Remove patch.crates-io

This FFI is managed by a dedicated crate to not impact build time for
the main crate.

To build the shared object:

  cargo build --release --package landlockconfig_ffi
  ls target/release/liblandlockconfig.so

The C/C++ header file is generated at build time and included in this
commit as a golden file.

Here are the exported functions:

  struct landlockconfig *landlockconfig_parse_json(int config_fd);
  struct landlockconfig *landlockconfig_parse_toml(int config_fd);
  int landlockconfig_build_ruleset(const struct landlockconfig *config);
  void landlockconfig_free(struct landlockconfig *config);

Update the CI:

- Run all tests on both packages.  The documentation is only built for
  landlockconfig because of name collision.
- Check if the ffi/c/landlockconfig.h golden file is synchronized with
  the source.
- Run the FFI example.

Closes: https://github.com/landlock-lsm/landlockconfig/issues/13
Signed-off-by: Mickaël Salaün <mic@digikod.net>