Fix handling of only one kind of access right
Commit 34aa832a4c93 ("Refactor Config to be idempotent") changed the way
a ruleset is populated, always making two calls to
Ruleset::handled_access(), one for each kind of access right.

Because the Landlock crate (rightfully) checks that we are not trying to
handle an empty set of access right, Config::build_ruleset() may return
an error if not both a filesystem and a network access rights are
handled.

Fix this issue by only calling Ruleset::handled_access() when it is
required, which restores the possibility to request only one kind of
access right.

Add a new CI test with verbose-write-tmp.json which only enforce
filesystem restrictions.

Signed-off-by: Mickaël Salaün <mic@digikod.net>