Add the ability to compose configurations
Every configurations can be composed with any other to get the union of
them in a safe and consistent way.  This might require to upgrade access
rights when they are not supported by one configuration to make a
consistent security policy.  This new operation is commutative.

Replace Variable::insert() with extend() to make it possible to compose
variable definitions.

Closes: https://github.com/landlock-lsm/landlockconfig/issues/20
Signed-off-by: Mickaël Salaün <mic@digikod.net>