Automatically infer handled access rights
This can be useful to have smaller configuration files, but explicitly
handling a set of access rights will always be useful because it
identifies a set of actions which are denied by default.

Additionally, the parser should limit error cases as much as possible.
One way to achieve that is to automatically complete the known
handled access rights

Rename test_inconsistent_handled_access() to
test_infer_mixed_handled_and_rule() and update the compared value to
reflect the infered configuration.

Add a new micro-write-tmp.toml example and test it with the CI.

Closes: https://github.com/landlock-lsm/landlockconfig/issues/12
Signed-off-by: Mickaël Salaün <mic@digikod.net>