Require non-empty configuration
It is not required to specify a ruleset anymore, which it possible
thanks to the handled access right inference, but at least one root
element must be set: ruleset, pathBeneath, or netPort.

Add a new NonEmptyStruct type to make sure the wrapped object has at
least one field set.  This makes sense when all fields of a struct are
optional (and can then all be None), but at least one of them must be
set.

Add more tests related the previous commit to check empty arrays.  It is
now easier because we don't need to have a ruleset entry, but only
pathBeneath or netPort.

Add a new micro-write-tmp.toml example and test it with the CI.

Signed-off-by: Mickaël Salaün <mic@digikod.net>