Use cargo-c for FFI
Generate a pkg-config file and other artifacts.

Synchronize the FFI crate version with the main one.

Update the C example to call cargo cinstall and build on top of the
installed files.

Keep c/landlockconfig.h to also use it as a golden file, and add a link
to the repository.

Update CI to install cargo-c and run the test target.

Update README.md with instructions to install cargo-c and how to use it.

Signed-off-by: Mickaël Salaün <mic@digikod.net>