Add parsing for simple variable with literal values
A variable is a set of values, which can be empty.  This is required to
fit well with the upcoming composability feature.

For now, a variable only stores a set of literal strings.  Future
version will add other types (e.g. environment variables).  It is
allowed to only declare a variable without value.

This change only stores the parsed variables from JSON/TOML.  The next
commit will interpret them.

Add a new tests_variable.rs file to test everything related to
variables.

Signed-off-by: Mickaël Salaün <mic@digikod.net>