Extrapolate file paths with variables
Variable names used in path_beneath's parent paths are identified with
the "${" prefix and the "}" suffix e.g., "${var}" .

Variables' name are checked and can only contain ASCII alphanumeric
characters and '_', but must start with a letter.  This is checked when
declaring and using variables.

The "$" character is escaped with another "$", which means that "$$"
translates to "$".  This makes it possible to write "$${not a variable".

Add a dedicated Variables type to easily extrapolate template strings.

Add TemplateString and related TemplateStringVisitor types to parse a
string with a finite state machine.

The new VecStringIterator type contains all the logic to extend a
template string by creating the Cartesian product of all possible values
in a deterministic order.

Test helpers are extended to be able to check against the JSON schema
validator's errors.  This is needed because we cannot define the format
of variable use in JSON schema.

Add a new examples/micro-var.toml configuration file to easily test
with:

  cargo run --example sandboxer -- --toml examples/micro-var.toml --debug date

Add this example to the README.md file.

Link: https://github.com/landlock-lsm/landlockconfig/issues/24
Signed-off-by: Mickaël Salaün <mic@digikod.net>