Cache toolchains and builds
To make it simple, cache the toolchain (Cargo and Rustup) and the latest
build per toolchain (MSRV and stable).  Each cache key is unique to make
sure we create a new cache each time, but the latest cache is still used
each time.

This is mostly useful for the FFI builds (from 6+ minutes to 7 seconds)
Running the full CI can now takes less than a minute.

Signed-off-by: Mickaël Salaün <mic@digikod.net>