Add abi.* access groups
Add the abi.all for all access types, plus abi.read_execute and
abi.read_write for filesystems access type.

Dealing with vN.* access groups does not scale well, is very verbose,
and can be confusing.  Add a new way to handle groups of access rights
by setting an ABI for a whole configuration file.  This new approach is
flexible enough and simpler, which helps users and maintenance..

The behavior of composition is unchanged except that Config may have an
explicit ABI value.

This feature is now possible thanks to the composition feature and it is
similar to the use of a local variable.

For instance, if we define `abi = 4`, this represents the highest
version of the Landlock ABI in a configuration file (for used groups),
which will replace the hardcoded v4 uses.  This is convenient to update
configurations to newest Landlock features by only updating one line
instead of all use of vN.

Example:

  abi = 4

  [[ruleset]]
  handled_access_fs = ["abi.all"]

  [[path_beneath]]
  allowed_access = ["abi.read_execute"]
  parent = ["/usr"]

Add new ValueAccess enums and AbiGroup trait to factor out common code.
Replace get_fs_read_execute() and get_fs_read_write() with
AbiGroupFs::ReadExecute and AbiGroupFs::ReadWrite.

Add tests_abi.rs to check ABI evaluation.

Bump the Landlock crate to get the Debug, Ord, and Eq implementations
for ABI.

Signed-off-by: Mickaël Salaün <mic@digikod.net>