Widely share Rust built caches
Do not use the commit ID to avoid almost duplicated caches.  GitHub CI
caches are only shared for the same PR and can be imported from the main
branch.

GitHub CI cache is limited to 10 GB, a Rust build cache is ~1150 MB, and
an Arch Linux container cache is ~960 MB.

Signed-off-by: Mickaël Salaün <mic@digikod.net>